Fix up tests for #271
As part of it's work the load_modules_from_config() file imports modules.
However those modules rely on opsdroid being fully instantiated, and that is
not the case in these unit tests.  So some mocking was necessary to isolate
the test, and only test the successful parsing of the configuration file.