Handle missing section details in load_modules_from_config (#271)
* Handle missing section details in load_modules_from_config

If the user has a 'databases:' section in their configuration, but nothing
listed as part of it, the opsdroid.loader.load_modules_from_config() would
call _load_modules with the modules parameter set to None, which would cause
a problem when trying to iterate it on line 223.

* Add some failing tests for PR#271

I found that I was getting two different errors with with the
load_modules_from_config() call when the minimal.yaml file was in
different states. Hence I've added a minimal_2.yaml to cover both error
conditions.

* Fix bug where configuration node has no children elements.

If the configuration has elements with no children (such as the "hello"
skill in the tests/configs/minimal.yaml file) then a yaml import gives you
that element as a string, not a dict. This change makes
Loader._load_modules() deal with that sensibly.

* Fix up tests for #271

As part of it's work the load_modules_from_config() file imports modules.
However those modules rely on opsdroid being fully instantiated, and that is
not the case in these unit tests.  So some mocking was necessary to isolate
the test, and only test the successful parsing of the configuration file.

* Minor change to satisfy pylint.