Clean exit for Matrix connector
When a user was pressing ctrl+c the Matrix connector was catching the
`CancelledError` as part of the catchall `except` and just logging it to
the console. The next (and subsequent) loops through would raise a
`RuntimeError('Session is closed')` resulting in an infinite loop
logging the same error to the console over and over.

By correctly catching the `CancelledError` we can ensure that the Matrix
connector exits cleanly.