Suppress Matrix timeout errors in console (fixes #835)
The Matrix connector will often receive a timeout error when doing a
sync request. The catch-all except would just dump the entire HTML of
the timeout response to the console. This output can be quite voluminous
and not that useful. By checking for know timeout error codes we can
suppress the useless HTML output and just log the occurrence at the info
level for those that are interested to know when timeouts occur.

In addition to the standard `504 Gateway Timeout` error we also check
for `522 Connection timed out` and `524 A Timeout Error` which are
timeout errors specific to Cloudflare. See:

  https://support.cloudflare.com/hc/en-us/articles/115003011431/

We check for these Cloudflare errors because the largest Matrix host
(matrix.org) is fronted by Cloudflare.