Updated loader.py - added dots
Aside from 285, added dots to the end of all '_LOGGER.debug(' prints as well as on the end of some 'LOGGER.info(' like 340 as well as on '_LOGGER.error(' like 664. 

The most common fix included the use of ', "."' although some dots were added to the end of some sentences directly, when there was no risk of the dot affecting the code.