Core changes without new tests or machinery (#1682)
* Core changes without new tests or machiners

* Comment problematic test

* satisfy lint

* Fix core signal handling test

Once task creation moved from load() to start(), several things started
broke in test_signals (web_server = None had no attribute start(),
start() failed because there were no configured skills, etc).

Also, because test_signals was using threading, any exceptions were NOT visible
in the main thread. The thread died, and the test passed. Plus, the assertions
were essentially ignored because UnitTest uses exceptions to mark tests as
failed, but again, the exceptions were not visible in the main thread.

Plus, eventloop.add_signal_handlers() needs to be run in the main thread
according to the docs. It's unclear to me if creating Opsdroid() in the main
thread was sufficient to have add_signal_handlers run in the main thread.

So, drop threading in test_signals, mock opsdroid.start with something that
still runs the task loop and stop mocking sys.exit. Instead, use pytest.raises
to ensure that SystemExit was raised (by sys.exit).

* Isolate test_signals

For whatever reason, sending SIGHUP is causing pytest to exit in some test
scenarios. But running the test alone works fine. So, just always run it
separately.

* skip un-migrated test

The test_watchdog test has not been refactored to handle the refactored core.
So, skip it. Plus, as it was marked xfail, at least one import was missing, so it
was already failing on all platforms. Fix that by importing contextlib.

* Uncomment test

Might have intermittent failures, so leave a comment in case it still fails

* Increase test coverage by uncommenting test

* Move pytest -v opt to setup.cfg

Co-authored-by: Jacob Floyd <cognifloyd@gmail.com>
Co-authored-by: Stuart Mumford <stuart@cadair.com>