Fixes #1691
Also, while I was here, I noticed that when a message cannot be parsed---for example if it's missing the key 'text'---the log message gives just the exception's msg but not its type, as in:

Unable to parse message 'text'
when the exception is KeyError: text. I changed the error log to an exception log which includes the traceback and the content of the message in question. If that's not desired feel free to force-push away the commit cc1fafc.

Finally, I was confused by the bot-name configuration parameter---isn't the name of the bot set by gitter? I tentatively made a change in 1ffadae that ignores the bot-name configuration parameter and removes it from the documented example. Instead, the bot name is obtained from the username reported back by gitter. Feel free to blow that away if I'm off track.