Fixed an errant entry into the MongoDB documentation
When I was originally working on the Mongo functionality, I had
proposed a change for the "default_collection_name" in the main
configuration, as well as the ability to specify a "collection_name"
per call. However, upon discussion with opsdroid devs, we decided
instead to implement a context manager similar to another database
such as not to change the API. I made the required changes, but
it looks as though I missed removing one of the documentation changes
related to this. So I'm correcting that now.